update: address pr comments
