+++ src/components/ThreeDEditor.jsx
@@ -36,7 +36,7 @@ import { LABEL_TYPES, MEASUREMENT_MODES } from "../enums";
 import {
     defaultMeasurementsSettings,
     MeasurementSettingsHandler,
-} from "../mixins/measurements/all";
+} from "../mixins/measurements/MeasurementSettingsHandler";
 import settings from "../settings";
 import IconsToolbar from "./IconsToolbar";
 import ParametersMenu from "./ParametersMenu";
